+++ frb_codegen/src/generator/dart/ty_boxed.rs
@@ -1,7 +1,43 @@
 use crate::generator::dart::ty::TypeDartGeneratorTrait;
+use crate::ir::IrType::{Primitive, StructRef};
 use crate::ir::*;
 
 #[derive(Debug, Clone)]
 pub struct TypeBoxedGenerator(IrTypeBoxed);
 
-impl TypeDartGeneratorTrait for TypeBoxedGenerator {}
+impl TypeDartGeneratorTrait for TypeBoxedGenerator {
+    fn api2wire_body(&self) -> String {
+        match &*self.0.inner {
+            Primitive(_) => {
+                format!("return inner.new_{}(raw);", ty.safe_ident())
+            }
+            inner => {
+                format!(
+                    "final ptr = inner.new_{}();
+                    _api_fill_to_wire_{}(raw, ptr.ref);
+                    return ptr;",
+                    ty.safe_ident(),
+                    inner.safe_ident(),
+                )
+            }
+        }
+    }
+
+    fn api_fill_to_wire_body(&self) -> String {
+        if !matches!(*self.0.inner, Primitive(_)) {
+            format!(
+                " _api_fill_to_wire_{}(apiObj, wireObj.ref);",
+                boxed.inner.safe_ident()
+            )
+        } else {
+            "".to_string()
+        }
+    }
+
+    fn wire2api_body(&self) -> String {
+        match &*self.0.inner {
+            StructRef(inner) => format!("return _wire2api_{}(raw);", inner.safe_ident()),
+            _ => gen_wire2api_simple_type_cast(&ty.dart_api_type()),
+        }
+    }
+}
